Continue development to support ID collisions for different label types.

+++ graphquery/core/db/query/query.h
@@ -37,10 +37,10 @@ namespace graphquery::database::query
         *        that person’s friends, where the creation date of the
         *        message is less than the max date parameter. (limited to 20)
         *
-        *  \param _person_id uint64_t _person_id - ID of person
-        *  \param _max_date uint64_t - max date threshold
+        *  \param _person_id uint32_t _person_id - ID of person
+        *  \param _max_date uint32_t - max date threshold
         ***************************************************************/
-        [[nodiscard]] std::vector<std::map<std::string, std::string>> interaction_complex_2(uint64_t _person_id, uint64_t _max_date) const noexcept;
+        [[nodiscard]] std::vector<std::map<std::string, std::string>> interaction_complex_2(uint32_t _person_id, uint32_t _max_date) const noexcept;
 
         /****************************************************************
          ** \brief Provided a person ID, find the most recent comments that
@@ -48,27 +48,27 @@ namespace graphquery::database::query
          *         replies, and return each person and their reply comment.
          *         (limited to 20)
          *
-         *  \param _person_id uint64_t - ID of person
+         *  \param _person_id uint32_t - ID of person
          ***************************************************************/
-        [[nodiscard]] std::vector<std::map<std::string, std::string>> interation_complex_8(uint64_t _person_id) const noexcept;
+        [[nodiscard]] std::vector<std::map<std::string, std::string>> interation_complex_8(uint32_t _person_id) const noexcept;
 
         /****************************************************************
          ** \brief Creates an edge between two nodes (person, post),
          *         with passed parameters ID parameters.
          *
-         *  \param _person_id uint64_t - ID of person
-         *  \param _post_id uint64_t - ID of post
+         *  \param _person_id uint32_t - ID of person
+         *  \param _post_id uint32_t - ID of post
          ***************************************************************/
-        void interation_update_2(uint64_t _person_id, uint64_t _post_id) const noexcept;
+        void interation_update_2(uint32_t _person_id, uint32_t _post_id) const noexcept;
 
         /****************************************************************
         ** \brief Creates an edge between two nodes (person, person),
         *         with passed parameters ID parameters.
         *
-        *  \param _src_person_id uint64_t - ID of src person
-        *  \param _dst_person_id uint64_t - ID of dst person
+        *  \param _src_person_id uint32_t - ID of src person
+        *  \param _dst_person_id uint32_t - ID of dst person
         ***************************************************************/
-        void interation_update_8(uint64_t _src_person_id, uint64_t _dst_person_id) const noexcept;
+        void interation_update_8(uint32_t _src_person_id, uint32_t _dst_person_id) const noexcept;
 
         /****************************************************************
          ** \brief Given a person and post ID, delete the "LIKES" relationship
@@ -78,17 +78,17 @@ namespace graphquery::database::query
          *  \param _person_id - ID of person
          *  \param _post_id - ID of post
          ***************************************************************/
-        inline void interation_delete_2(uint64_t _person_id, uint64_t _post_id) const noexcept;
+        inline void interation_delete_2(uint32_t _person_id, uint32_t _post_id) const noexcept;
 
         /****************************************************************
          ** \brief Given two person IDs, delete the "KNOWS" relationship
          *         between the two person nodes, and then return the count
          *         of the deleted relationships.
          *
-         *  \param _src_person_id uint64_t - ID of person
+         *  \param _src_person_id uint32_t - ID of person
          *  \param _dst_person_id - ID of person
          ***************************************************************/
-        inline void interation_delete_8(uint64_t _src_person_id, uint64_t _dst_person_id) const noexcept;
+        inline void interation_delete_8(uint32_t _src_person_id, uint32_t _dst_person_id) const noexcept;
 
         /****************************************************************
         ** \brief Given a person ID, returns the 10 most recent messages
@@ -98,7 +98,7 @@ namespace graphquery::database::query
         *
         *  \param _person_id - ID of person
 s        ***************************************************************/
-        void interation_short_2(uint64_t _person_id) const noexcept;
+        void interation_short_2(uint32_t _person_id) const noexcept;
 
         /****************************************************************
          ** \brief Given the message ID, returns comments made on a specific
@@ -107,9 +107,9 @@ s        ***************************************************************/
          *         It then orders the results by comment creation date and
          *         the ID of the comment authors.
          *
-         *  \param _message_id uint64_t - ID of message
+         *  \param _message_id uint32_t - ID of message
          ***************************************************************/
-        void interation_short_7(uint64_t _message_id) const noexcept;
+        void interation_short_7(uint32_t _message_id) const noexcept;
 
       private:
         [[nodiscard]] storage::ILPGModel * get_graph() const noexcept;
